requirements.txt
discord.py[voice]==2.3.2
flask[async]==3.0.2
nest_asyncio==1.6.0
requests==2.31.0
selenium==4.18.1
